// apps/web/src/lib/api/generated/model/visitResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits, workflow, billing, products, reviews
 * OpenAPI spec version: 1.0.0
 */
import type { VisitResponseStatus } from './visitResponseStatus';

export interface VisitResponse {
  id?: number;
  petId?: number;
  petName?: string;
  petBreed?: string;
  petBirthDate?: string;
  vetId?: number;
  customerUserId?: string;
  ownerName?: string;
  ownerPhone?: string;
  scheduledAt?: string;
  status?: VisitResponseStatus;
  reason?: string;
  diagnosis?: string;
  treatment?: string;
  fee?: number;
  processInstanceKey?: number;
}